app/requirements.txt
fastapi
python-multipart
uvicorn
gunicorn
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow-gpu==1.15.0
keras==2.0.8
opencv-python
h5py==2.10.0
imgaug
IPython[all]